feat(books, shelves, users): enhance filtering and ordering capabilities in views and serializers

+++ shelves/views.py
@@ -3,7 +3,7 @@ from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
-from rest_framework.filters import SearchFilter
+from rest_framework.filters import SearchFilter, OrderingFilter
 from .models import Shelf
 from users.models import CustomUser
 from .serializers import *
@@ -22,25 +22,18 @@ class ShelfList(ListCreateAPIView):
         return {'request': self.request}
 
 
-""" @api_view(['GET'])
-def shelves_list(request):
-    shelves = Shelf.objects.select_related(
-        'user').prefetch_related('shelfbook_set__book').all()
-    serializer = ShelfSerializer(
-        shelves, many=True, context={'request': request})
-    return Response(serializer.data)
- """
-
-
 class UserShelfList(ListCreateAPIView):
-    filter_backends = [DjangoFilterBackend, SearchFilter]
+    filter_backends = [DjangoFilterBackend, SearchFilter, OrderingFilter]
     filterset_class = ShelfFilter
     search_fields = ['name']
+    ordering_fields = ['shelfbook', 'name', 'shelfbook__book',
+                       'shelfbook__date_added', 'shelfbook__date_finished',
+                       ]
 
     def get_queryset(self):
         user_id = self.kwargs['user_id']
         user = get_object_or_404(CustomUser, pk=user_id)
-        return Shelf.objects.select_related('user').prefetch_related('shelfbook_set__book').filter(user=user)
+        return Shelf.objects.select_related('user').prefetch_related('shelfbook_set__book__genres').filter(user=user)
 
     def get_serializer_class(self):
         if self.request.method == 'POST':
@@ -59,6 +52,7 @@ class UserShelfList(ListCreateAPIView):
 class ShelfDetails(RetrieveUpdateDestroyAPIView):
     queryset = Shelf.objects.select_related(
         'user').prefetch_related('shelfbook_set__book').all()
+    filter_backends = [DjangoFilterBackend, SearchFilter, OrderingFilter]
 
     def get_serializer_class(self):
         if self.request.method in ['PUT', 'PATCH']:
@@ -67,23 +61,6 @@ class ShelfDetails(RetrieveUpdateDestroyAPIView):
 
     def get_serializer_context(self):
         return {'request': self.request}
-
-
-""" @api_view(['GET', 'PUT', 'PATCH'])
-def shelf_details(request, pk):
-    shelf = get_object_or_404(Shelf, pk=pk, user=request.user)
-
-    if request.method == 'GET':
-        serializer = ShelfSerializer(shelf, context={'request': request})
-        return Response(serializer.data)
-
-    elif request.method in ['PUT', 'PATCH']:
-        partial = request.method == 'PATCH'
-        serializer = ShelfCreateSerializer(
-            shelf, data=request.data, partial=partial)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data) """
 
 
 @api_view(['POST', 'DELETE', 'PUT', 'PATCH'])
@@ -100,7 +77,7 @@ def book_to_shelf(request, pk, book_id=None):
         serializer.save(shelf=shelf)
         return Response(serializer.data, status=201)
 
-    elif request.method == 'DELETE':
+    elif request.method == 'DELETE' and book_id:
         shelf_book = get_object_or_404(
             ShelfBook, shelf=shelf, book=book_id)
         shelf_book.delete()
